improve brevo rake tasks (more verbose)

+++ lib/tasks/brevo.rake
@@ -14,8 +14,16 @@ namespace :brevo do
   task list_inbound_email_events: :environment do
     action_mailbox_password = Rails.application.credentials.action_mailbox.ingress_password
     api_instance = BrevoRuby::InboundParsingApi.new
-    api_instance.get_inbound_email_events.events[0..10].each do |event|
-      pp api_instance.get_inbound_email_events_by_uuid(event.uuid)
+    api_instance.get_inbound_email_events.events[0..3].each_with_index do |event, index|
+      detailed_event = api_instance.get_inbound_email_events_by_uuid(event.uuid)
+      logs = detailed_event.logs.map { |log| "#{log.type} (#{log.date})" }
+      puts "--- EVENT ##{index + 1} ---"
+      puts "Subject: #{detailed_event.subject}"
+      puts "To: #{detailed_event.recipient}"
+      puts "Received at: #{detailed_event.received_at}"
+      puts "Attachments?: #{detailed_event.attachments.size > 0}"
+      puts "Logs: #{logs.join(' -> ')}"
+      puts
     end
   end
 
@@ -42,13 +50,14 @@ namespace :brevo do
     end
   end
 
-  desc 'Delete all the local Brevo webhooks'
-  task delete_local_webhooks: :environment do
+  desc 'Delete the local Brevo webhook'
+  task delete_local_webhook: :environment do
     api_instance = BrevoRuby::WebhooksApi.new
     result = api_instance.get_webhooks(type: 'inbound')
     result.webhooks.each do |webhook|
+      next if webhook[:domain] != ENV['INBOUND_REPLY_EMAIL_DOMAIN']
       api_instance.delete_webhook(webhook[:id])
-      puts "Webhook ##{webhook[:id]} deleted ✅"
+      puts "Webhook ##{webhook[:id]} (#{webhook[:domain]}) deleted ✅"
     end
   end
 end
